requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.1.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tavily-python>=0.3.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
